Add SipHash13
SipHash13 is secure enough to be used as a hash function in a hash table.
SipTreeHash13 is comparable in performance to HighwayTreeHash on
long strings.

It has no practically usable attack as a MAC.
But it is quite close to "theoretically usable" distinguisher.
So that, it's use as a secure MAC is not recommended.