Adjust cl/559350994 to use const void* to work with new Clang
See
https://github.com/llvm/llvm-project/pull/67766#issuecomment-1741485234
and https://lists.llvm.org/pipermail/llvm-dev/2016-January/094012.html .
An unaligned pointer has an unspecified value per C++ expr.static.cast,
in which the compiler is permitted to disallow it being passed to
memcpy. Clang -fsanitize=alignment recently becomes stricter (memcpy is
now instrumented) in part to complement AddressSanitizer detection.

There are quite a few LoadUnaligned overloads in the code base that
tickles this unspecified behavior, but they are benign for Clang.

PiperOrigin-RevId: 578755824